DAOS-11239 control: Stub SPDK library interface in go unit tests
Provide stubbing for the SPDK C library interface so that go does not
link to the actual library unless the spdk build flag is provided.

Features: control
Allow-unstable: true
Required-githooks: true
