+++ src/control/SConscript
@@ -50,6 +50,7 @@ def get_build_tags(benv):
     if is_firmware_mgmt_build(benv):
         tags.append("firmware")
     tags.append("ucx")
+    tags.append("spdk")
     return "-tags {}".format(','.join(tags))
