+++ src/control/lib/ipmctl/nvm_types.go
@@ -39,48 +39,6 @@ func (p PartNumber) String() string {
 	return bytes2String(p[:])
 }
 
-// DeviceDiscovery struct represents Go equivalent of C.struct_device_discovery
-// from nvm_management.h (NVM API) as reported by "go tool cgo -godefs nvm.go"
-type DeviceDiscovery struct {
-	All_properties_populated uint8
-	Pad_cgo_0                [3]byte
-	Device_handle            [4]byte
-	Physical_id              uint16
-	Vendor_id                uint16
-	Device_id                uint16
-	Revision_id              uint16
-	Channel_pos              uint16
-	Channel_id               uint16
-	Memory_controller_id     uint16
-	Socket_id                uint16
-	Node_controller_id       uint16
-	Pad_cgo_1                [2]byte
-	Memory_type              uint32
-	Dimm_sku                 uint32
-	Manufacturer             [2]uint8
-	Serial_number            [4]uint8
-	Subsystem_vendor_id      uint16
-	Subsystem_device_id      uint16
-	Subsystem_revision_id    uint16
-	Manufacturing_info_valid uint8
-	Manufacturing_location   uint8
-	Manufacturing_date       uint16
-	Part_number              PartNumber
-	Fw_revision              Version
-	Fw_api_version           Version
-	Pad_cgo_2                [5]byte
-	Capacity                 uint64
-	Interface_format_codes   [9]uint16
-	Security_capabilities    _Ctype_struct_device_security_capabilities
-	Device_capabilities      _Ctype_struct_device_capabilities
-	Uid                      DeviceUID
-	Lock_state               uint32
-	Manageability            uint32
-	Controller_revision_id   uint16
-	Reserved                 [48]uint8
-	Pad_cgo_3                [6]byte
-}
-
 // PMemRegionType represents PMem region type.
 type PMemRegionType uint32
